backend/requirements.txt
asgiref==3.9.2
dj-database-url==3.0.1
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
